fix an error

+++ plexusnet/architecture.py
@@ -592,7 +592,7 @@ class PlexusNet():
         for i in range(length):
             if self.lanewise_augmentation:
                 if i in self.ApplyLaneForAugmentation:
-                    x = self.data_augmentation(x)
+                    x = self.data_augmentation(tf.cast(x, dtype=tf.float32)
             vb = self.Spider_Node(x, initial_filter, compression,depth, kernel_regularizer, counter=i, type_of_block=type_of_block, initial_image=initial_image)
         for j, layer in enumerate(vb.layers):
             layer._name = f"C_{i}_{layer.name}"
